# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend server for Lastnight",
        "title": "Lastnight API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "skndash96",
            "email": "dashskndash@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Authenticates a user with email and password",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login endpoint",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "delete": {
                "description": "Logs out the current user by clearing the authentication token cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout endpoint",
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Registers a new user and returns an authentication token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Signup endpoint",
                "parameters": [
                    {
                        "description": "Register credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "description": "Returns the health status of the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Healthcheck endpoint",
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/teams": {
            "get": {
                "description": "Get the user's teams list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Get Teams",
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/teams/default": {
            "post": {
                "description": "Join the default team for the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Join Default Team",
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/teams/{teamID}/tags": {
            "get": {
                "description": "Get the tags of a team",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Get Tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "New Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/teams/{teamID}/tags/{tagID}": {
            "put": {
                "description": "Update a tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Update Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Delete Tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/teams/{teamID}/tags/{tagID}/values": {
            "get": {
                "description": "Get the values of a tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Get Tag Values",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tag value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Create Tag Value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/teams/{teamID}/tags/{tagID}/values/{valueID}": {
            "delete": {
                "description": "Delete a tag value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Delete Tag Value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Value ID",
                        "name": "valueID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.GetTeamsByUserIDRow": {
            "type": "object",
            "properties": {
                "membership_id": {
                    "type": "integer"
                },
                "team_domain": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "team_name": {
                    "type": "string"
                },
                "user_joined_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "user_role": {
                    "$ref": "#/definitions/db.TeamUserRole"
                }
            }
        },
        "db.Tag": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "data_type": {
                    "$ref": "#/definitions/db.TagDataType"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "db.TagDataType": {
            "type": "string",
            "enum": [
                "string",
                "number",
                "boolean"
            ],
            "x-enum-varnames": [
                "TagDataTypeString",
                "TagDataTypeNumber",
                "TagDataTypeBoolean"
            ]
        },
        "db.TagValue": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "id": {
                    "type": "integer"
                },
                "tag_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "db.Team": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "db.TeamUserRole": {
            "type": "string",
            "enum": [
                "member",
                "mod"
            ],
            "x-enum-varnames": [
                "TeamUserRoleMember",
                "TeamUserRoleMod"
            ]
        },
        "dto.CreateTagRequest": {
            "type": "object",
            "required": [
                "data_type",
                "name",
                "teamID"
            ],
            "properties": {
                "data_type": {
                    "maxLength": 100,
                    "minLength": 2,
                    "allOf": [
                        {
                            "$ref": "#/definitions/db.TagDataType"
                        }
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "teamID": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateTagResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Tag"
                }
            }
        },
        "dto.CreateTagValueResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.TagValue"
                }
            }
        },
        "dto.DeleteTagResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Tag"
                }
            }
        },
        "dto.DeleteTagValueResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.TagValue"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email is required field"
                }
            }
        },
        "dto.GetTagValuesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.TagValue"
                    }
                }
            }
        },
        "dto.GetTagsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Tag"
                    }
                }
            }
        },
        "dto.GetTeamsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GetTeamsByUserIDRow"
                    }
                }
            }
        },
        "dto.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "dto.JoinTeamResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Team"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object"
        },
        "dto.LogoutResponse": {
            "type": "object"
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.RegisterResponse": {
            "type": "object"
        },
        "dto.UpdateTagRequest": {
            "type": "object",
            "required": [
                "name",
                "tagID",
                "teamID"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "tagID": {
                    "type": "integer"
                },
                "teamID": {
                    "type": "integer"
                }
            }
        },
        "dto.UpdateTagResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Tag"
                }
            }
        },
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    }
}