# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type TagDataType string

const (
	TagDataTypeString  TagDataType = "string"
	TagDataTypeNumber  TagDataType = "number"
	TagDataTypeBoolean TagDataType = "boolean"
)

func (e *TagDataType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TagDataType(s)
	case string:
		*e = TagDataType(s)
	default:
		return fmt.Errorf("unsupported scan type for TagDataType: %T", src)
	}
	return nil
}

type NullTagDataType struct {
	TagDataType TagDataType `json:"tag_data_type"`
	Valid       bool        `json:"valid"` // Valid is true if TagDataType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTagDataType) Scan(value interface{}) error {
	if value == nil {
		ns.TagDataType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TagDataType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTagDataType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TagDataType), nil
}

type TeamUserRole string

const (
	TeamUserRoleMember TeamUserRole = "member"
	TeamUserRoleMod    TeamUserRole = "mod"
)

func (e *TeamUserRole) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TeamUserRole(s)
	case string:
		*e = TeamUserRole(s)
	default:
		return fmt.Errorf("unsupported scan type for TeamUserRole: %T", src)
	}
	return nil
}

type NullTeamUserRole struct {
	TeamUserRole TeamUserRole `json:"team_user_role"`
	Valid        bool         `json:"valid"` // Valid is true if TeamUserRole is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTeamUserRole) Scan(value interface{}) error {
	if value == nil {
		ns.TeamUserRole, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TeamUserRole.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTeamUserRole) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TeamUserRole), nil
}

type Account struct {
	ID                int32            `json:"id"`
	UserID            int32            `json:"user_id"`
	Provider          string           `json:"provider"`
	ProviderAccountID string           `json:"provider_account_id"`
	Password          []byte           `json:"password"`
	CreatedAt         pgtype.Timestamp `json:"created_at"`
	UpdatedAt         pgtype.Timestamp `json:"updated_at"`
}

type Tag struct {
	ID        int32            `json:"id"`
	TeamID    int32            `json:"team_id"`
	Name      string           `json:"name"`
	DataType  TagDataType      `json:"data_type"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type TagValue struct {
	ID        int32            `json:"id"`
	TagID     int32            `json:"tag_id"`
	Value     string           `json:"value"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type Team struct {
	ID        int32            `json:"id"`
	Name      string           `json:"name"`
	Domain    string           `json:"domain"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type TeamMemberTag struct {
	ID               int32            `json:"id"`
	TeamMembershipID int32            `json:"team_membership_id"`
	TagID            int32            `json:"tag_id"`
	TagValueID       int32            `json:"tag_value_id"`
	CreatedAt        pgtype.Timestamp `json:"created_at"`
}

type TeamMembership struct {
	ID       int32            `json:"id"`
	TeamID   int32            `json:"team_id"`
	UserID   int32            `json:"user_id"`
	Role     TeamUserRole     `json:"role"`
	JoinedAt pgtype.Timestamp `json:"joined_at"`
}

type User struct {
	ID        int32            `json:"id"`
	Email     string           `json:"email"`
	Name      string           `json:"name"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}